requirements.txt

Django==5.2.8
django-debug-toolbar==6.1.0
django-extensions==4.1
pillow==12.0.0
django-thumbnails=0.8.0
django-ckeditor=6.7.3